Move signature checks and more
